backend/requirements.txt
redis>=5.0.0
celery>=5.3.0
neo4j>=5.14.0
openai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
